Restructure ReactDiagram classes into src/diagram
