+++ src/clients/LocalClient.ts
@@ -1,5 +1,5 @@
 import { Server } from '@data-story-org/core';
-import DiagramModel from '../DiagramModel';
+import DiagramModel from '../diagram/models/DiagramModel';
 import ClientInterface from './ClientInterface';
 
 const context = {
