+++ src/diagram/factories/NodeModelFactory.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { AbstractReactFactory } from '@projectstorm/react-canvas-core';
-import NodeModel from './NodeModel';
-import NodeWidget from './components/widgets/Node/';
-import CommentNodeWidget from './components/CommentNodeWidget';
+import NodeModel from '../models/NodeModel';
+import NodeWidget from '../../components/widgets/Node/';
+import CommentNodeWidget from '../../components/CommentNodeWidget';
 
 export default class NodeModelFactory extends AbstractReactFactory {
   constructor() {
